rename data-page and create workspace-base class
