+++ e2e/tests/cohorts/cohorts-rename.spec.ts
@@ -1,6 +1,6 @@
 import Link from 'app/element/link';
 import CohortBuildPage, {FieldSelector} from 'app/page/cohort-build-page';
-import DataPage, {TabLabelAlias} from 'app/page/data-page';
+import WorkspaceDataPage, {TabLabelAlias} from 'app/page/workspace-data-page';
 import {findWorkspace, signIn} from 'utils/test-utils';
 import {waitForNumericalString, waitForText} from 'utils/waits-utils';
 import DataResourceCard from 'app/component/data-resource-card';
@@ -31,7 +31,7 @@ describe('User can create, modify, rename and delete Cohort', () => {
     const workspaceCard = await findWorkspace(page);
     await workspaceCard.clickWorkspaceName();
 
-    const dataPage = new DataPage(page);
+    const dataPage = new WorkspaceDataPage(page);
     // Click Add Cohorts button in Data page.
     await dataPage.getAddCohortsButton().then((butn) => butn.clickAndWait());
