+++ e2e/tests/cohorts/cohorts-ui.spec.ts
@@ -1,9 +1,9 @@
-import {PhysicalMeasurementsCriteria} from 'app/page/cohort-criteria-modal';
 import CohortBuildPage from 'app/page/cohort-build-page';
-import DataPage, {TabLabelAlias} from 'app/page/data-page';
+import {PhysicalMeasurementsCriteria} from 'app/page/cohort-criteria-modal';
+import WorkspaceAboutPage from 'app/page/workspace-about-page';
+import WorkspaceDataPage, {TabLabelAlias} from 'app/page/workspace-data-page';
 import {findWorkspace, signIn} from 'utils/test-utils';
 import {waitForText} from 'utils/waits-utils';
-import WorkspaceAboutPage from 'app/page/workspace-about-page';
 
 describe('Cohorts UI tests', () => {
 
@@ -22,7 +22,7 @@ describe('Cohorts UI tests', () => {
     await workspaceCard.clickWorkspaceName();
 
     // Wait for the Data page.
-    const dataPage = new DataPage(page);
+    const dataPage = new WorkspaceDataPage(page);
 
     const addCohortsButton = await dataPage.getAddCohortsButton();
     await addCohortsButton.clickAndWait();
@@ -65,7 +65,7 @@ describe('Cohorts UI tests', () => {
     // Check ABOUT tab is open
     const aboutPage = new WorkspaceAboutPage(page);
     await aboutPage.waitForLoad();
-    const isOpen = await aboutPage.isOpen();
+    const isOpen = await aboutPage.isOpen(TabLabelAlias.About);
     expect(isOpen).toBe(true);
   });
