+++ e2e/tests/notebook/notebook-save.spec.ts
@@ -1,5 +1,5 @@
 import DataResourceCard, {CardType} from 'app/component/data-resource-card';
-import DataPage, {TabLabelAlias} from 'app/page/data-page';
+import WorkspaceDataPage, {TabLabelAlias} from 'app/page/workspace-data-page';
 import WorkspaceAnalysisPage from 'app/page/workspace-analysis-page';
 import {Language} from 'app/text-labels';
 import {makeRandomName} from 'utils/str-utils';
@@ -20,7 +20,7 @@ describe('Jupyter Notebook tests in Python language', () => {
     const workspaceCard = await findWorkspace(page);
     await workspaceCard.clickWorkspaceName();
 
-    const dataPage = new DataPage(page);
+    const dataPage = new WorkspaceDataPage(page);
     await dataPage.openTab(TabLabelAlias.Analysis);
 
     const notebookName = makeRandomName('py-notebook');
